pyproject.toml
[project]
name = "04-practise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.8",
    "openai-agents>=0.2.5",
]
